Moved server into one file
